@@ -5,7 +5,7 @@ namespace CesarBmx.CryptoWatcher.Api.Configuration
 {
     public static class MiddlewareConfig
     {
-        public static IApplicationBuilder ConfigureErrorHandling(this IApplicationBuilder app)
+        public static IApplicationBuilder ConfigureMiddleware(this IApplicationBuilder app)
         {
             app.ConfigureSharedMiddleware();
 
