@@ -8,10 +8,10 @@ namespace CesarBmx.CryptoWatcher.Api.Configuration
 {
     public static class SerilogConfig
     {
-        public static ILoggerFactory ConfigureSerilog(this IApplicationBuilder app, ILoggerFactory logger, IConfiguration configuration)
+        public static ILoggerFactory ConfigureSerilog(this ILoggerFactory logger, IConfiguration configuration)
         {
-            app.ConfigureSharedSerilog(logger, Assembly.GetExecutingAssembly(), configuration);
-            
+            logger.ConfigureSharedSerilog(configuration, Assembly.GetExecutingAssembly());
+
             return logger;
         }
     }
